fix compilation error from lottie-react dependency: `ReferenceError: document is not defined`
the release of Node v21 (in 2023) added a [`navigator` API](https://nodejs.org/api/globals.html#navigator)
(a lightweight version of `window.navigator`) by the same name. unfortunately,
[`lottie-react`](https://github.com/Gamote/lottie-react) imports
[`lottie-web`](https://github.com/airbnb/lottie-web), which at the
time had client-side guards that checked for
`typeof navigator !== "undefined"` instead of for `window`.

```
ReferenceError: document is not defined
    at createTag
(node_modules/lottie-web/build/player/lottie.js:30:5)
```

it was first reported in 2023:

- https://github.com/airbnb/lottie-web/issues/3047
- https://github.com/airbnb/lottie-web/issues/2739

it was fixed the following year, airbnb/lottie-web#3133, but the maintainer of `lottie-react` still
hasn't released a new version that bumps the `lottie-web` dependency. it's either fork the project
or this not-terrible workaround. regardless, it would be a wise performance improvement for
lightening page loads.